+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/geogrid/GeoHashGridAggregator.java
@@ -44,10 +44,10 @@ public class GeoHashGridAggregator extends BucketsAggregator {
 
     private final int requiredSize;
     private final int shardSize;
-    private final GeoGridAggregationBuilder.CellIdSource valuesSource;
+    private final CellIdSource valuesSource;
     private final LongHash bucketOrds;
 
-    GeoHashGridAggregator(String name, AggregatorFactories factories, GeoGridAggregationBuilder.CellIdSource valuesSource,
+    GeoHashGridAggregator(String name, AggregatorFactories factories, CellIdSource valuesSource,
             int requiredSize, int shardSize, SearchContext aggregationContext, Aggregator parent,
             List<PipelineAggregator> pipelineAggregators, Map<String, Object> metaData) throws IOException {
         super(name, factories, aggregationContext, parent, pipelineAggregators, metaData);
@@ -96,7 +96,7 @@ public class GeoHashGridAggregator extends BucketsAggregator {
     }
 
     // private impl that stores a bucket ord. This allows for computing the aggregations lazily.
-    static class OrdinalBucket extends InternalGeoHashGrid.Bucket {
+    static class OrdinalBucket extends GeoGridBucket {
 
         long bucketOrd;
 
@@ -125,7 +125,7 @@ public class GeoHashGridAggregator extends BucketsAggregator {
             spare = (OrdinalBucket) ordered.insertWithOverflow(spare);
         }
 
-        final InternalGeoHashGrid.Bucket[] list = new InternalGeoHashGrid.Bucket[ordered.size()];
+        final GeoGridBucket[] list = new GeoGridBucket[ordered.size()];
         for (int i = ordered.size() - 1; i >= 0; --i) {
             final OrdinalBucket bucket = (OrdinalBucket) ordered.pop();
             bucket.aggregations = bucketAggregations(bucket.bucketOrd);
